fix for norm and separate parser.c to several files
