+++ srcs/lexer/lexer_token_checker.c
@@ -6,7 +6,7 @@
 /*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 14:14:38 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/11/03 05:21:15 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/11/22 04:46:55 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,5 +49,3 @@ bool	is_operator(char *input)
 	}
 	return (false);
 }
-
-
