+++ srcs/lexer/lexer_token_quote_checker.c
@@ -6,7 +6,7 @@
 /*   By: tamatsuu <tamatsuu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/03 05:21:17 by tamatsuu          #+#    #+#             */
-/*   Updated: 2024/11/03 05:23:01 by tamatsuu         ###   ########.fr       */
+/*   Updated: 2024/11/22 04:47:05 by tamatsuu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
